Extract ID scalar implmentation
This change moves the ID scalar implementation out of `graphql.go` and
into its own file `id.go` for consistency with the Time scalar
implementation.
